refactor: extract remarkPlugin to own file
